ANDR-5: id стало Long по аналогии с другими классами

+++ feature/interview-trainer/impl/src/main/java/ru/yeahub/interview_trainer/impl/createQuiz/presentation/CreateQuizState.kt
@@ -13,8 +13,8 @@ sealed interface CreateQuizState {
     ) : CreateQuizState {
         @Immutable
         data class VoSpecialization(
-            val id: Int,
-            val title: String
+            val id: Long,
+            val title: String,
         )
     }
